refactor(subagents): separate execution rendering and config
